requirements.txt
streamlit
crewai
crewai-tools
google-generativeai
python-dotenv
requests
pydantic